refactor: PostControllerTest.java를 테스트하는 기능에 따라 분리
